requirements/runtime.txt
Django>=2.2,<5.1
celery-redbeat==2.2.0
